requirements.txt
argon2-cffi==19.1.0
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
coverage==4.5.3
Django==2.2.1
djangorestframework==3.9.3
docutils==0.14
geoip2==2.9.0
httpie==1.0.2
idna==2.8
importlib-metadata==0.15
Jinja2==2.10.1
langdetect==1.0.7
MarkupSafe==1.1.1
maxminddb==1.4.1
more-itertools==7.0.0
nltk==3.4.1
numpy==1.16.3
Pillow==6.0.0
pluggy==0.12.0
py==1.8.0
pycparser==2.19
Pygments==2.4.0
pytest==4.5.0
pytest-cov==2.7.1
pytest-django==3.4.8
python-memcached==1.59
pytz==2019.1
PyYAML==5.1
requests==2.21.0
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
tblib==1.4.0
textblob==0.15.3
textblob-fr==0.2.0
urllib3==1.24.3
wcwidth==0.1.7
zipp==0.5.1
